Remove custom python override
It turns out the python override is fixing the wrong problem. All module
not found issue comes from the fact that `PATH` is overriden and wrong
python3 binary is being used when a program is called from fusesoc.

The change to `PATH` is caused by the python program wrapper, but in
this case it's not necessary for fusesoc. Simply stopping wrapping
fusesoc is enough to fix the issue.