spike: bump to head of 'ibex-cosim'
Since ibex#2259, ibex dv in ibex/master now requires spike from the head of the
'riscv-isa-sim/ibex-cosim' branch. A new release/tag may be cut in the future,
but for now any updates should just track this branch.

This branch now contains a commit for the patch
"fesvr-fix-compilation-with-gcc-13.patch", so remove it from the override.

Related:
lowRISC/ibex#2183,lowRISC/riscv-isa-sim#25
lowRISC/ibex#2259