Don't touch when binding if file exists in buildFHSEnvOverlay
buildFHSEnvOverlay provides a helper function bind() which can be used to
create mounts as part of its init script. This can bind to any arbitrary point
in the filesystem, including over the top of existing files or directories.

If you bind to a read-only filesystem, the touch command will fail, and
generate noise on stderr. This is irrelevant since the subsequent mount will
then succeed regardless.

To avoid this, we can just avoid the touch command if the path we are binding to
already exists.